refacotor: OptionList 에서 focus 로직 처리하도록 분리
