+++ packages/wow-ui/src/components/DropDown/context/DropDownContext.ts
@@ -1,7 +1,7 @@
 import { createContext } from "react";
 
-import type useDropDownState from "@/hooks/useDropDownState";
-import useSafeContext from "@/hooks/useSafeContext";
+import type useDropDownState from "../../../hooks/useDropDownState";
+import useSafeContext from "../../../hooks/useSafeContext";
 
 export const DropDownContext = createContext<ReturnType<
   typeof useDropDownState
